+++ packages/eslint-plugin-orbit-components/src/rules/buttonHasTitle.ts
@@ -1,7 +1,7 @@
 import isOrbitComponent from "../utils/isOrbitComponent";
 import detectOriginalOrbitName from "../utils/detectOriginalOrbitName";
 import { Rule } from "eslint";
-import { types as t } from "@babel/core";
+import * as t from "@babel/types";
 
 export default {
   create: (context: Rule.RuleContext) => {