+++ packages/orbit-components/src/Seat/__tests__/index.test.js
@@ -6,13 +6,13 @@ import Seat from "../index";
 import { SIZE_OPTIONS, TYPES } from "../consts";
 
 const height = {
-  [SIZE_OPTIONS.SMALL]: "57px",
-  [SIZE_OPTIONS.MEDIUM]: "65px",
+  [SIZE_OPTIONS.SMALL]: "36px",
+  [SIZE_OPTIONS.MEDIUM]: "46px",
 };
 
 const width = {
-  [SIZE_OPTIONS.SMALL]: "40px",
-  [SIZE_OPTIONS.MEDIUM]: "53px",
+  [SIZE_OPTIONS.SMALL]: "34px",
+  [SIZE_OPTIONS.MEDIUM]: "47px",
 };
 
 describe("Seat", () => {