+++ packages/orbit-components/src/Seat/index.js
@@ -2,25 +2,37 @@
 import styled from "styled-components";
 import * as React from "react";
 
+import Stack from "../Stack";
+import Text from "../Text";
 import randomID from "../utils/randomID";
 import defaultTheme from "../defaultTheme";
 import SeatItem from "./SeatItem";
-import { SIZE_OPTIONS } from "./consts";
+import { SIZE_OPTIONS, TYPES } from "./consts";
 
 import type { Props } from "./index";
 
-const getSize = ({ size }) => {
+const WrappedText = styled.div`
+  padding-right: ${({ selected, size }) =>
+    selected && (size === SIZE_OPTIONS.SMALL ? "6px" : "7px")};
+`;
+
+const getSize = ({ size, selected }) => {
   const height = {
-    [SIZE_OPTIONS.SMALL]: "57px",
-    [SIZE_OPTIONS.MEDIUM]: "65px",
+    [SIZE_OPTIONS.SMALL]: selected ? "43px" : "36px",
+    [SIZE_OPTIONS.MEDIUM]: selected ? "54px" : "46px",
   };
 
   const width = {
-    [SIZE_OPTIONS.SMALL]: "40px",
-    [SIZE_OPTIONS.MEDIUM]: "53px",
+    [SIZE_OPTIONS.SMALL]: selected ? "40px" : "34px",
+    [SIZE_OPTIONS.MEDIUM]: selected ? "54px" : "47px",
   };
 
   return `width: ${width[size]}; height: ${height[size]};`;
+};
+
+const resolveViewBoxSize = (size, selected) => {
+  if (selected) return size === SIZE_OPTIONS.SMALL ? "0 0 40 43" : "0 0 54 54";
+  return size === SIZE_OPTIONS.SMALL ? "0 0 34 36" : "0 0 47 46";
 };
 
 const StyledSeat = styled.svg`
@@ -48,17 +60,28 @@ const Seat = ({
   const descrId = React.useMemo(() => randomID("descr"), []);
 
   return (
-    <StyledSeat
-      viewBox={size === SIZE_OPTIONS.SMALL ? "0 0 40 57" : "0 0 53 65"}
-      aria-labelledby={`${titleId} ${descrId}`}
-      size={size}
-      role="img"
-      data-test={dataTest}
-    >
-      <title id={titleId}>{title}</title>
-      <desc id={descrId}>{description}</desc>
-      <SeatItem type={type} selected={selected} size={size} price={price} label={label} />
-    </StyledSeat>
+    <Stack inline grow={false} spacing="XXXSmall" direction="column" align="center">
+      <StyledSeat
+        viewBox={resolveViewBoxSize(size, selected)}
+        aria-labelledby={`${titleId} ${descrId}`}
+        size={size}
+        fill="none"
+        role="img"
+        selected={selected}
+        data-test={dataTest}
+      >
+        <title id={titleId}>{title}</title>
+        <desc id={descrId}>{description}</desc>
+        <SeatItem type={type} selected={selected} size={size} price={price} label={label} />
+      </StyledSeat>
+      {price && !(selected && type === TYPES.UNAVAILABLE) && (
+        <WrappedText selected={selected} size={size}>
+          <Text size="small" type="secondary">
+            {price}
+          </Text>
+        </WrappedText>
+      )}
+    </Stack>
   );
 };