+++ packages/orbit-components/src/Tag/index.js
@@ -3,10 +3,11 @@ import * as React from "react";
 import styled, { css } from "styled-components";
 
 import defaultTheme from "../defaultTheme";
-import { rtlSpacing, left, right } from "../utils/rtl";
+import { rtlSpacing, left } from "../utils/rtl";
 import CloseCircle from "../icons/CloseCircle";
 import { SIZES, STATES } from "./consts";
 import KEY_CODE_MAP from "../common/keyMaps";
+import resolveColor from "./helpers/resolveColor";
 
 import type { Props } from "./index";
 
@@ -19,45 +20,76 @@ const getFontSize = ({ theme, size }) => {
   return tokens[size];
 };
 
-const getBackgroundColor = state => ({ selected, theme }) => {
+const getBackgroundColor = state => () => {
   const states = {
-    [STATES.DEFAULT]: selected ? theme.orbit.paletteBlueLightHover : theme.orbit.paletteCloudDark,
-    [STATES.HOVER]: selected
-      ? theme.orbit.paletteBlueLightActive
-      : theme.orbit.paletteCloudNormalHover,
-    [STATES.ACTIVE]: selected
-      ? theme.orbit.paletteBlueLightActive
-      : theme.orbit.paletteCloudNormalHover,
+    [STATES.DEFAULT]: resolveColor({
+      selected: "paletteBlueNormal",
+      removable: "paletteBlueLight",
+      normal: "paletteCloudDark",
+    }),
+    [STATES.HOVER]: resolveColor({
+      selected: "paletteBlueNormalHover",
+      removable: "paletteBlueLightHover",
+      normal: "paletteCloudNormalHover",
+    }),
+    [STATES.ACTIVE]: resolveColor({
+      selected: "paletteBlueNormalActive",
+      removable: "paletteBlueLightActive",
+      normal: "paletteCloudNormalHover",
+    }),
   };
   return states[state];
 };
 
-export const StyledTag = styled.div`
-  font-family: ${({ theme }) => theme.orbit.fontFamily};
-  color: ${({ theme, selected }) =>
-    selected ? theme.orbit.paletteBlueDarker : theme.orbit.colorTextTag};
-  background: ${getBackgroundColor(STATES.DEFAULT)};
-  display: inline-flex;
+const CloseContainer = styled.div`
+  display: flex;
+  margin-${left}: 8px;
+  opacity: 0.5;
+  color: ${resolveColor({
+    selected: "paletteWhite",
+    removable: "paletteBlueDarker",
+    normal: "paletteInkLink",
+  })};
+  cursor: ${({ actionable }) => actionable && `pointer`};
+  transition: all ${({ theme }) => theme.orbit.durationFast} ease-in-out;
 
-  box-sizing: border-box;
-  justify-content: center;
-  align-items: center;
-  font-size: ${getFontSize};
-  font-weight: ${({ theme }) => theme.orbit.fontWeightMedium};
-  border-radius: ${({ theme }) => theme.orbit.borderRadiusNormal};
-  box-shadow: ${({ theme, selected }) =>
-    !selected && `inset 0 0 0 1px ${theme.orbit.borderColorTag}`};
-  padding: ${({ theme }) => rtlSpacing(theme.orbit.paddingTag)};
-  transition: color ${({ theme }) => theme.orbit.durationFast} ease-in-out,
-    box-shadow ${({ theme }) => theme.orbit.durationFast} ease-in-out,
-    background ${({ theme }) => theme.orbit.durationFast} ease-in-out;
-
-  &:focus {
-    outline: 0;
+  &:active {
+    color: ${({ theme, selected }) =>
+      selected ? theme.orbit.paletteWhite : theme.orbit.paletteBlueDarker};
   }
+`;
 
-  ${({ actionable }) =>
-    actionable &&
+CloseContainer.defaultProps = {
+  theme: defaultTheme,
+};
+
+export const StyledTag = styled.div`
+  ${({ theme, actionable }) => css`
+    font-family: ${theme.orbit.fontFamily};
+    color: ${resolveColor({
+      selected: "paletteWhite",
+      removable: "paletteBlueDarker",
+      normal: "colorTextTag",
+    })};
+    background: ${getBackgroundColor(STATES.DEFAULT)};
+    display: inline-flex;
+
+    box-sizing: border-box;
+    justify-content: center;
+    align-items: center;
+    font-size: ${getFontSize};
+    font-weight: ${theme.orbit.fontWeightMedium};
+    border-radius: ${theme.orbit.borderRadiusNormal};
+    padding: ${rtlSpacing(theme.orbit.paddingTag)};
+    transition: color ${theme.orbit.durationFast} ease-in-out,
+      box-shadow ${theme.orbit.durationFast} ease-in-out,
+      background ${theme.orbit.durationFast} ease-in-out;
+
+    &:focus {
+      outline: 0;
+    }
+
+    ${actionable &&
     css`
       cursor: pointer;
 
@@ -67,55 +99,26 @@ export const StyledTag = styled.div`
       }
 
       &:active {
+        ${CloseContainer} {
+          opacity: 1;
+        }
         background: ${getBackgroundColor(STATES.ACTIVE)};
         box-shadow: none;
       }
 
       &:focus {
+        ${CloseContainer} {
+          opacity: 1;
+        }
         background: ${getBackgroundColor(STATES.HOVER)};
         box-shadow: none;
         outline: 0;
       }
     `};
+  `}
 `;
 
 StyledTag.defaultProps = {
-  theme: defaultTheme,
-};
-
-const IconContainer = styled.div`
-  display: flex;
-  margin-${left}: 8px;
-
-  svg {
-    height: ${({ theme }) => theme.orbit.widthIconSmall};
-    width: ${({ theme }) => theme.orbit.heightIconSmall};
-  }
-`;
-
-IconContainer.defaultProps = {
-  theme: defaultTheme,
-};
-
-const CloseContainer = styled.div`
-  display: flex;
-  margin-${right}: 8px;
-  color: ${({ theme, selected }) =>
-    selected ? theme.orbit.paletteBlueDarker : theme.orbit.paletteInkLight};
-  cursor: pointer;
-  transition: color ${({ theme }) => theme.orbit.durationFast} ease-in-out;
-
-  &:hover {
-    color: ${({ theme, selected }) =>
-      selected ? theme.orbit.paletteBlueDarker : theme.orbit.paletteInkLight};
-  }
-  &:active {
-    color: ${({ theme, selected }) =>
-      selected ? theme.orbit.paletteBlueDarker : theme.orbit.paletteInkLight};
-  }
-`;
-
-CloseContainer.defaultProps = {
   theme: defaultTheme,
 };
 
@@ -124,10 +127,13 @@ const StyledClose = styled.div`
   border-radius: 100%;
 
   &:focus {
+    ${CloseContainer} {
+      opacity: 1;
+    }
     outline: none;
     box-shadow: 0 0 0 2px
-      ${({ theme, selected }) =>
-        selected ? theme.orbit.paletteBlueDarker : theme.orbit.paletteBlueDarker};
+      ${({ theme, selected, removable }) =>
+        selected && !removable ? theme.orbit.paletteWhite : theme.orbit.paletteBlueDarker};
   }
 `;
 
@@ -135,57 +141,56 @@ StyledClose.defaultProps = {
   theme: defaultTheme,
 };
 
-const buttonClickEmulation = (ev, callback) => {
+const buttonClickEmulation = callback => (ev?: SyntheticKeyboardEvent<HTMLButtonElement>) => {
   if (ev && ev.keyCode === KEY_CODE_MAP.SPACE) {
     ev.preventDefault();
     if (callback) callback();
-  } else if (ev.keyCode === KEY_CODE_MAP.ENTER) {
+  } else if (ev && ev.keyCode === KEY_CODE_MAP.ENTER) {
     if (callback) callback();
   }
 };
 
-const Tag = (props: Props) => {
-  const { icon, selected, children, size = SIZES.NORMAL, onClick, onRemove, dataTest } = props;
+const Tag = React.forwardRef<Props, HTMLDivElement>((props, ref) => {
+  const { selected, children, size = SIZES.NORMAL, onClick, onRemove, dataTest } = props;
   return (
     <StyledTag
       actionable={onClick || onRemove}
       data-test={dataTest}
       size={size}
+      ref={ref}
       onClick={onClick}
       removable={!!onRemove}
       selected={selected}
-      icon={!!icon}
-      tabIndex="0"
+      tabIndex={(onClick || onRemove) && "0"}
       role="button"
-      onKeyDown={ev => buttonClickEmulation(ev, onClick)}
+      onKeyDown={buttonClickEmulation(onClick)}
     >
+      {children}
       {!!onRemove && (
         <CloseContainer
           selected={selected}
+          removable={!!onRemove}
           onClick={ev => {
             ev.stopPropagation();
-            if (onRemove) {
-              onRemove();
-            }
+            if (onRemove) onRemove();
           }}
         >
           <StyledClose
             tabIndex="0"
+            selected={selected}
             aria-label="close"
             role="button"
             onKeyDown={ev => {
               ev.stopPropagation();
-              buttonClickEmulation(ev, onRemove);
+              buttonClickEmulation(onRemove);
             }}
           >
             <CloseCircle size="small" />
           </StyledClose>
         </CloseContainer>
       )}
-      {children}
-      {icon && <IconContainer>{icon}</IconContainer>}
     </StyledTag>
   );
-};
+});
 
 export default Tag;