+++ packages/eslint-plugin-orbit-components/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.3](https://github.com/kiwicom/orbit/compare/eslint-plugin-orbit-components@0.2.2...eslint-plugin-orbit-components@0.2.3) (2021-01-22)
+
+
+### Bug Fixes
+
+* stop relying on a transient dependency ([#2675](https://github.com/kiwicom/orbit/issues/2675)) ([4895971](https://github.com/kiwicom/orbit/commit/48959716c4ea209b94ad3efec4af68a92605910c)), closes [#2660](https://github.com/kiwicom/orbit/issues/2660) [#2674](https://github.com/kiwicom/orbit/issues/2674)
+
+
+
+
+
 ## [0.2.2](https://github.com/kiwicom/orbit/compare/eslint-plugin-orbit-components@0.2.1...eslint-plugin-orbit-components@0.2.2) (2021-01-13)